requirements.txt
robotframework
mock
pyautogui
tk
opencv-python
scikit-image
matplotlib
# pyqt5 is only needed to use skimage.viewer.ImageViewer(refImage) while debugging
# pyqt5